Decomposing alignment function to alerts module
